Backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-jose
passlib[bcrypt]
python-dotenv
redis
httpx
motor
email_validator
